Move amount formatter to Report PDF
